+++ src/CompilerState.ts
@@ -58,7 +58,12 @@ export class CompilerState {
 		return id;
 	}
 
+	// TODO: Remove the unused Truthy ones
 	public currentConditionalContext: string = "";
+	public currentTruthyContext: string = "";
+	public topTruthyNode?: ts.BinaryExpression;
+	public hasTruthyContextBeenUsed?: boolean;
+	public binaryRhsExpressions = new Map<ts.Expression, ts.Expression>();
 	private precedingStatementContexts = new Array<PrecedingStatementContext>();
 
 	public getCurrentPrecedingStatementContext(node: ts.Node) {